\docType{package}
\name{ImageArtR-package}
\alias{ImageArtR}
\alias{ImageArtR-package}
\title{ImageArtR: Turn Images into Artistic Graphics}
\description{
ImageArtR reads images from files, URLs, \code{magick-image} objects, or
\code{imager::cimg} objects, then converts them into circle packing art,
mosaics, outlines, sketches, cartoons, and color palettes.
}
\seealso{
\code{\link{image_circle_art}}, \code{\link{image_mosaic_art}},
\code{\link{image_outline}}, \code{\link{image_sketch}}
}
