% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/artify.R
\name{list_art_styles}
\alias{list_art_styles}
\title{List available art styles}
\usage{
list_art_styles()
}
\value{
A tibble with style metadata.
}
\description{
List available art styles
}
\examples{
list_art_styles()
}
