% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pop-art.R
\name{list_pop_art_palettes}
\alias{list_pop_art_palettes}
\title{List built-in pop art palettes}
\usage{
list_pop_art_palettes()
}
\value{
A tibble with palette names and color vectors.
}
\description{
List built-in pop art palettes
}
\examples{
list_pop_art_palettes()
}
