% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image-class.R
\name{plot.image_art}
\alias{plot.image_art}
\title{Plot an image art object}
\usage{
\method{plot}{image_art}(x, ...)
}
\arguments{
\item{x}{An \code{image_art} object.}

\item{...}{Unused.}
}
\value{
A \code{ggplot2} object, invisibly.
}
\description{
Plot an image art object
}
