% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image-class.R
\name{print.image_art}
\alias{print.image_art}
\title{Print an image art object}
\usage{
\method{print}{image_art}(x, ...)
}
\arguments{
\item{x}{An \code{image_art} object.}

\item{...}{Unused.}
}
\value{
The input object, invisibly.
}
\description{
Print an image art object
}
