% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ascii-art.R
\name{print_ascii_art}
\alias{print_ascii_art}
\title{Print ASCII art}
\usage{
print_ascii_art(x, ...)
}
\arguments{
\item{x}{An ASCII \code{image_art} object, character vector, or text string.}

\item{...}{Unused.}
}
\value{
The ASCII lines, invisibly.
}
\description{
Print ASCII art
}
